Implement constructor invocation.
We can now define classes, invoke methods, and set and get attributes (and __call, __get and __set are called when necessary), and we can instantiate classes; however, if the class definition contains definition of attributes, those attributes are not yet added to the instantiated objects. Not 100% sure what the correct way to do this is; perhaps modify the constructor?


git-svn-id: http://phc.googlecode.com/svn/trunk@2296 bbcf20a9-8526-0410-90e2-9b3d3f589b1d